+++ src/main/java/run/halo/weapp/endpoint/ConfigEndpoint.java
@@ -33,6 +33,7 @@ public class ConfigEndpoint implements CustomEndpoint {
                     String requestId = ErrorHandler.newRequestId();
                     return Mono.fromCallable(() -> PublicConfig.from(
                             settingsService.site(),
+                            settingsService.features(),
                             settingsService.comment(),
                             settingsService.announcement(),
                             settingsService.client()))